delete user

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/service/UserService.java
@@ -67,6 +67,9 @@ public class UserService {
     }
 
     public void deleteUser(long userId) {
+        if (userRepository.findById(userId) == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "The user with the given id does not exist. Therefore, the user could not be deleted!");
+        }
         User userToDelete = getUser(userId);
         userRepository.delete(userToDelete);
     }
